+++ debug-draw-3d-plugin/src/debugdraw3d/debug_draw_manager.h
@@ -22,13 +22,13 @@ class DebugDrawManager;
 class _DD3D_PhysicsWatcher : public Node {
 	GDCLASS(_DD3D_PhysicsWatcher, Node)
 protected:
-	DebugDrawManager *root_node;
-	static void _bind_methods(){};
+	DebugDrawManager* root_node;
+	static void _bind_methods() {};
 
 public:
-	void init(DebugDrawManager *p_root);
+	void init(DebugDrawManager* p_root);
 
-	virtual void _physics_process(float p_delta) override;
+	virtual void _physics_process(double p_delta) override;
 };
 #endif
 
@@ -42,11 +42,14 @@ public:
  *
  * `debug_draw_3d/settings/initial_debug_state` sets the initial debugging state.
  *
- * `debug_draw_3d/settings/common/DebugDrawManager_singleton_aliases` sets aliases for DebugDrawManager to be registered as additional singletons.
+ * `debug_draw_3d/settings/common/DebugDrawManager_singleton_aliases` sets aliases for DebugDrawManager to be registered
+ * as additional singletons.
  *
- * `debug_draw_3d/settings/common/DebugDraw2D_singleton_aliases` sets aliases for DebugDraw2D to be registered as additional singletons.
+ * `debug_draw_3d/settings/common/DebugDraw2D_singleton_aliases` sets aliases for DebugDraw2D to be registered as
+ * additional singletons.
  *
- * `debug_draw_3d/settings/common/DebugDraw3D_singleton_aliases` sets aliases for DebugDraw3D to be registered as additional singletons.
+ * `debug_draw_3d/settings/common/DebugDraw3D_singleton_aliases` sets aliases for DebugDraw3D to be registered as
+ * additional singletons.
  *
  * Using these aliases you can reference singletons with shorter words:
  *
@@ -67,34 +70,34 @@ private:
 	friend _DD3D_PhysicsWatcher;
 #endif
 
-	static DebugDrawManager *singleton;
+	static DebugDrawManager* singleton;
 
 	String root_settings_section;
-	const static char *s_initial_state;
-	const static char *s_manager_aliases;
-	const static char *s_dd2d_aliases;
-	const static char *s_dd3d_aliases;
+	const static char* s_initial_state;
+	const static char* s_manager_aliases;
+	const static char* s_dd2d_aliases;
+	const static char* s_dd3d_aliases;
 
 	double log_flush_time = 0;
 	bool debug_enabled = true;
 	bool is_closing = false;
 	bool is_current_scene_is_null = true;
 
-	DebugDraw2D *debug_draw_2d_singleton = nullptr;
-	DebugDraw3D *debug_draw_3d_singleton = nullptr;
+	DebugDraw2D* debug_draw_2d_singleton = nullptr;
+	DebugDraw3D* debug_draw_3d_singleton = nullptr;
 
 	TypedArray<StringName> manager_aliases;
 	TypedArray<StringName> dd2d_aliases;
 	TypedArray<StringName> dd3d_aliases;
 
-	Node *_get_current_scene();
+	Node* _get_current_scene();
 	void _connect_scene_changed();
 	void _on_scene_changed(bool p_is_scene_null);
 
 	void _integrate_into_engine();
 
-	void _register_singleton_aliases(const TypedArray<StringName> &p_names, Object *p_instance);
-	void _unregister_singleton_aliases(const TypedArray<StringName> &p_names);
+	void _register_singleton_aliases(const TypedArray<StringName>& p_names, Object* p_instance);
+	void _unregister_singleton_aliases(const TypedArray<StringName>& p_names);
 
 #ifdef TOOLS_ENABLED
 	void _try_to_update_cs_bindings();
@@ -114,27 +117,35 @@ public:
 	};
 
 private:
-	static Object *default_arg_obj;
+	static Object* default_arg_obj;
 
 	// Test regular arguments
-	void api_test1(Variant, Object *, bool, int, float, String, StringName, NodePath){};
+	void api_test1(Variant, Object*, bool, int, float, String, StringName, NodePath) {};
 	void api_test2(Color, Vector2, Vector2i, Vector3, Vector3i, Vector4, Vector4i, Rect2, Rect2i){};
 	void api_test3(Transform2D, Transform3D, Plane, Quaternion, AABB, Basis, Projection){};
 	void api_test4(RID, Callable, Signal, Dictionary, Array){};
-	void api_test5(PackedByteArray, PackedInt32Array, PackedInt64Array, PackedFloat32Array, PackedFloat64Array, PackedStringArray, PackedVector2Array, PackedVector3Array, PackedColorArray){};
+	void api_test5(PackedByteArray, PackedInt32Array, PackedInt64Array, PackedFloat32Array, PackedFloat64Array,
+			PackedStringArray, PackedVector2Array, PackedVector3Array, PackedColorArray){};
 	// Test with default arguments
-	Variant api_test6(Object *, Variant, Variant, bool, int, DevTestEnum, float, String, StringName, NodePath) { return "test var"; };
-	Color api_test7(Color, Vector2, Vector2i, Vector3, Vector3i, Vector4, Vector4i, Rect2, Rect2i) { return Color(4, 3, 2, 1); };
-	DevTestEnum api_test8(Transform2D, Transform3D, Plane, Quaternion, AABB, Basis, Projection) { return (DevTestEnum)1; };
-	Object *api_test9(RID, Callable, Signal, Dictionary, Array) { return this; };
-	void api_test10(PackedByteArray, PackedInt32Array, PackedInt64Array, PackedFloat32Array, PackedFloat64Array, PackedStringArray, PackedVector2Array, PackedVector3Array, PackedColorArray){};
+	Variant api_test6(Object*, Variant, Variant, bool, int, DevTestEnum, float, String, StringName, NodePath) {
+		return "test var";
+	};
+	Color api_test7(Color, Vector2, Vector2i, Vector3, Vector3i, Vector4, Vector4i, Rect2, Rect2i) {
+		return Color(4, 3, 2, 1);
+	};
+	DevTestEnum api_test8(Transform2D, Transform3D, Plane, Quaternion, AABB, Basis, Projection) {
+		return (DevTestEnum)1;
+	};
+	Object* api_test9(RID, Callable, Signal, Dictionary, Array) { return this; };
+	void api_test10(PackedByteArray, PackedInt32Array, PackedInt64Array, PackedFloat32Array, PackedFloat64Array,
+			PackedStringArray, PackedVector2Array, PackedVector3Array, PackedColorArray){};
 
 public:
 #endif
 
 public:
 	/// @private
-	static const char *s_extension_unloading;
+	static const char* s_extension_unloading;
 
 	DebugDrawManager();
 	~DebugDrawManager();
@@ -142,9 +153,7 @@ public:
 	/**
 	 * Get singleton. Not available in GDScript.
 	 */
-	static DebugDrawManager *get_singleton() {
-		return singleton;
-	};
+	static DebugDrawManager* get_singleton() { return singleton; };
 
 #pragma region Exposed Methods
 	/**
@@ -167,11 +176,11 @@ public:
 	void deinit();
 
 	/// @private
-	virtual void _process(float p_delta) override;
+	virtual void _process(double p_delta) override;
 	/// @private
 	void _physics_process_start(float p_delta);
 	/// @private
-	virtual void _physics_process(float p_delta) override;
+	virtual void _physics_process(double p_delta) override;
 };
 
 #ifdef DEV_ENABLED
