+++ gameplay-plugin/include/components/inputcomponent.h
@@ -130,7 +130,7 @@ public:
 	static void _bind_methods();
 
 	void _notification(int what);
-	void _physics_process(float delta);
+	void _physics_process(double delta);
 	void _input(const Ref<InputEvent>& p_event);
 	void _unhandled_input(const Ref<InputEvent>& p_event);
